requirements.txt
streamlit==1.45.1
streamlit-extras==0.4.0
Pillow==9.0.1
